requirements.txt
Flask==3.0.2
firebase-admin==6.4.0
python-dotenv==1.0.1
requests==2.31.0
Werkzeug==3.0.1
pyrebase4==4.8.0 